Closes GH-0010 POM cleanup, remove unneeded repositories, remove/exclude commons-cli 2.0
iThe main POM file contains references to snapshot repositories which
are not needed to build Oozie.

The core and example POM have references to commons-cli 2.0 that are
not needed.

The POMs of hadoop artifacts used by Oozie have references to
commons-cli 2.0 which are not needed and they should be excluded.

From: Alejandro Abdelnur <tucu00@gmail.com>

git-svn-id: https://svn.apache.org/repos/asf/incubator/oozie/branches/2.2@1169090 13f79535-47bb-0310-9956-ffa450edef68