Closes GH-0262 WorkflowAppService should always inject Kerberos principal rules regardless of Kerberos/Simple auth
It seems that changes in Hadoop code require the principal/rules to be present even if using security simple.

As the rules do not seem to be generated automatically from Hadoop -default.xml file Oozie has to inject them.

The WorkflowAppService currently is injecting the rules only if Kerberos authentication is ON.

This injection should happen always to avoid exceptions with newer Hadoop code.

From: Alejandro Abdelnur <tucu00@gmail.com>

git-svn-id: https://svn.apache.org/repos/asf/incubator/oozie/branches/3.0@1169204 13f79535-47bb-0310-9956-ffa450edef68